requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
jupyter
requests
python-dotenv
